Make logger a global object.
